refactor(GuessStatus): 패키지 이동
